[FEAT] 통합테스트 구현 및 디스코드, 스케줄링 #45
